Added "NoDevice" to the list of Portaudio errors
This is the error that portaudio returns if you ask for the default
device, but the system doesn't have a default device.

Before this change, Portaudio returning NoDevice would result in a
panic. Now, it will result in the new error being returned.

I also updated the examples to avoid panicking, and rather print the
error message and exit gracefully.

Related: https://github.com/RustAudio/rust-portaudio/issues/142